Updates to Code/Showing data on pages
Had to slightly adjust the code in the api and backend files to be able to get the data to display on each page, although currently the whole table is displayed meaning that all the data from the memorialnames table is showing on the page at once. Will need to work on showing around 20 or 30 records at a time.